move core functionality to dedicated file
