+++ go.mod
@@ -1,3 +1,15 @@
 module github.com/masonbrott/sigilkv
 
 go 1.23.1
+
+require (
+	github.com/andybalholm/brotli v1.1.1 // indirect
+	github.com/fasthttp/router v1.5.4 // indirect
+	github.com/klauspost/compress v1.18.0 // indirect
+	github.com/qiangxue/fasthttp-routing v0.0.0-20160225050629-6ccdc2a18d87 // indirect
+	github.com/savsgio/gotils v0.0.0-20240704082632-aef3928b8a38 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasthttp v1.60.0 // indirect
+	go.uber.org/multierr v1.10.0 // indirect
+	go.uber.org/zap v1.27.0 // indirect
+)
